refactor: cleanup package-utils package
