+++ .eslintignore
@@ -7,7 +7,6 @@ packages/generators/lib
 packages/info/lib
 packages/init/lib
 packages/migrate/lib
-packages/package-utils/lib
 packages/serve/lib
 packages/utils/lib
 packages/webpack-scaffold/lib
