+++ packages/generators/__tests__/addon-generator.test.ts
@@ -1,4 +1,4 @@
-jest.setMock('@webpack-cli/package-utils', {
+jest.setMock('@webpack-cli/utils', {
     getPackageManager: jest.fn(),
 });
 
@@ -6,7 +6,7 @@ import fs from 'fs';
 import path from 'path';
 import mkdirp from 'mkdirp';
 import rimraf from 'rimraf';
-import { getPackageManager } from '@webpack-cli/package-utils';
+import { getPackageManager } from '@webpack-cli/utils';
 import addonGenerator from '../src/addon-generator';
 
 // TODO: enable after jest release
