+++ packages/generators/src/addon-generator.ts
@@ -2,8 +2,7 @@ import logger from 'webpack-cli/lib/utils/logger';
 import mkdirp from 'mkdirp';
 import path from 'path';
 import Generator from 'yeoman-generator';
-import { getPackageManager } from '@webpack-cli/package-utils';
-import { generatorCopy, generatorCopyTpl } from '@webpack-cli/utils';
+import { generatorCopy, generatorCopyTpl, getPackageManager } from '@webpack-cli/utils';
 
 /**
  * Creates a Yeoman Generator that generates a project conforming
