+++ packages/generators/src/init-generator.ts
@@ -2,7 +2,7 @@ import { blue, green, bold } from 'colorette';
 import logger from 'webpack-cli/lib/utils/logger';
 import logSymbols from 'log-symbols';
 import path from 'path';
-import { getPackageManager } from '@webpack-cli/package-utils';
+import { getPackageManager } from '@webpack-cli/utils';
 import { Confirm, Input, List } from '@webpack-cli/webpack-scaffold';
 
 import {
