--- packages/package-utils/CHANGELOG.md
@@ -1,34 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [1.0.1-rc.1](https://github.com/webpack/webpack-cli/compare/@webpack-cli/package-utils@1.0.1-alpha.4...@webpack-cli/package-utils@1.0.1-rc.1) (2020-10-06)
-
-### Bug Fixes
-
--   **utils:** respect package-lock.json ([#1375](https://github.com/webpack/webpack-cli/issues/1375)) ([ce8ec5a](https://github.com/webpack/webpack-cli/commit/ce8ec5a9f56ab5c1ce30742dced56dcbea237600))
--   **webpack-cli:** handle promise rejection with package installation ([#1284](https://github.com/webpack/webpack-cli/issues/1284)) ([eb1112e](https://github.com/webpack/webpack-cli/commit/eb1112edf05b0a1bc83dced0e83987e4f459174c))
--   check webpack installation before running cli ([#1827](https://github.com/webpack/webpack-cli/issues/1827)) ([be509fa](https://github.com/webpack/webpack-cli/commit/be509fac9a03e202e062229484bb10af7876968f))
--   declare default commands as optional peer dependency ([#1816](https://github.com/webpack/webpack-cli/issues/1816)) ([dcc75c8](https://github.com/webpack/webpack-cli/commit/dcc75c8b24d2d170a9274d80ecfb8b2329e4ad2e))
--   use appropriate exit codes ([#1755](https://github.com/webpack/webpack-cli/issues/1755)) ([83f73b0](https://github.com/webpack/webpack-cli/commit/83f73b056e224301b871bee5e9b7254e64e84e95))
-
-## [1.0.1-alpha.4](https://github.com/webpack/webpack-cli/compare/@webpack-cli/package-utils@1.0.1-alpha.3...@webpack-cli/package-utils@1.0.1-alpha.4) (2020-03-02)
-
-**Note:** Version bump only for package @webpack-cli/package-utils
-
-## [1.0.1-alpha.3](https://github.com/webpack/webpack-cli/compare/@webpack-cli/package-utils@1.0.1-alpha.2...@webpack-cli/package-utils@1.0.1-alpha.3) (2020-02-29)
-
-**Note:** Version bump only for package @webpack-cli/package-utils
-
-## [1.0.1-alpha.2](https://github.com/webpack/webpack-cli/compare/@webpack-cli/package-utils@1.0.1-alpha.1...@webpack-cli/package-utils@1.0.1-alpha.2) (2020-02-23)
-
-**Note:** Version bump only for package @webpack-cli/package-utils
-
-## [1.0.1-alpha.1](https://github.com/webpack/webpack-cli/compare/@webpack-cli/package-utils@1.0.1-alpha.0...@webpack-cli/package-utils@1.0.1-alpha.1) (2020-02-23)
-
-**Note:** Version bump only for package @webpack-cli/package-utils
-
-## 1.0.1-alpha.0 (2020-02-23)
-
-**Note:** Version bump only for package @webpack-cli/package-utils
