--- packages/package-utils/README.md
@@ -1,11 +1,0 @@
-# `@webpack-cli/package-utils`
-
-> TODO: description
-
-## Usage
-
-```
-const packageUtils = require('@webpack-cli/package-utils');
-
-// TODO: DEMONSTRATE API
-```
