--- packages/package-utils/package.json
@@ -1,57 +1,0 @@
-{
-  "name": "@webpack-cli/package-utils",
-  "version": "1.0.1-rc.1",
-  "description": "A module to help managing packages and modules inside webpack CLI",
-  "keywords": [
-    "webpack",
-    "webpack-cli",
-    "node",
-    "package",
-    "utils",
-    "npm",
-    "yarn"
-  ],
-  "author": "emanuele <my.burning@gmail.com>",
-  "homepage": "https://github.com/webpack/webpack-cli#readme",
-  "license": "MIT",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "directories": {
-    "test": "__tests__"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/webpack/webpack-cli.git"
-  },
-  "bugs": {
-    "url": "https://github.com/webpack/webpack-cli/issues"
-  },
-  "files": [
-    "lib"
-  ],
-  "dependencies": {
-    "colorette": "^1.2.1",
-    "cross-spawn": "^7.0.1",
-    "enquirer": "^2.3.4",
-    "execa": "^4.0.0",
-    "global-modules": "^2.0.0"
-  },
-  "peerDependenciesMeta": {
-    "webpack": {
-      "optional": true
-    },
-    "@webpack-cli/info": {
-      "optional": true
-    },
-    "@webpack-cli/init": {
-      "optional": true
-    },
-    "@webpack-cli/serve": {
-      "optional": true
-    }
-  },
-  "gitHead": "fb50f766851f500ca12867a2aa9de81fa6e368f9"
-}
