--- packages/package-utils/src/index.ts
@@ -1,2 +1,0 @@
-export * from './packageUtils';
-export * from './processUtils';
