+++ packages/utils/src/index.ts
@@ -9,3 +9,5 @@ export * from './run-prettier';
 export * from './scaffold';
 export * from './validate-identifier';
 export * from './prop-types';
+export * from './global-packages-path';
+export * from './get-package-manager';
