+++ packages/utils/tsconfig.json
@@ -4,6 +4,5 @@
     "outDir": "./lib",
     "rootDir": "./src"
   },
-  "include": ["./src"],
-  "references": [{ "path": "../package-utils" }]
+  "include": ["./src"]
 }
